Cleaning *General reorganization of files and functions
