+++ .gitignore
@@ -1,2 +1,5 @@
 venv/
 __pycache__/
+test.ipynb
+database.db
+.vscode/
