+++ layouts/constants.py
@@ -1,0 +1,2 @@
+degree_sign = u'\N{DEGREE SIGN}'
+month_list = ['Enero','Febrero','Marzo','Abril','Mayo','Junio','Julio','Agosto','Septiembre','Octubre','Noviembre','Diciembre']
